This commit will add a preliminary auto script.
There will now be a task package with classes for each task of the
auto scripting language (bundle, serial, wait, print, cargo, hatch,
climb, drive, follow/path, and turn) and an AutoCompiler class to
call each task depending on tokens for each in an auto mode file.

This is based on the code for the 2018 auto script and is largely
incomplete (path, drive, turn, cargo, and lift tasks are yet to be
defined) but will serve as a basis for future additions.

Also, the enums for each subsystem's modes will now be located in each
corresponding task file (removed from UrsaRobot).